Extract getInformationFromTestnode
