+++ src/testHelpers.ts
@@ -1,6 +1,7 @@
 import { sha256, toBytes } from 'viem';
 import { privateKeyToAccount, PrivateKeyAccount } from 'viem/accounts';
 import { config } from 'dotenv';
+import { execSync } from 'node:child_process';
 
 import { sanitizePrivateKey } from './utils';
 
@@ -55,4 +56,41 @@ export function getNitroTestnodePrivateKeyAccounts(): NitroTestNodePrivateKeyAcc
       privateKey: l3TokenBridgeDeployerPrivateKey,
     },
   };
+}
+
+type TestnodeInformation = {
+  rollup: `0x${string}`;
+  l3Rollup: `0x${string}`;
+  l3NativeToken: `0x${string}`;
+};
+
+export function getInformationFromTestnode(): TestnodeInformation {
+  const containers = [
+    'nitro_sequencer_1',
+    'nitro-sequencer-1',
+    'nitro-testnode-sequencer-1',
+    'nitro-testnode_sequencer_1',
+  ];
+
+  for (const container of containers) {
+    try {
+      const deploymentJson = JSON.parse(
+        execSync('docker exec ' + container + ' cat /config/deployment.json').toString(),
+      );
+
+      const l3DeploymentJson = JSON.parse(
+        execSync('docker exec ' + container + ' cat /config/l3deployment.json').toString(),
+      );
+
+      return {
+        rollup: deploymentJson['rollup'],
+        l3Rollup: l3DeploymentJson['rollup'],
+        l3NativeToken: l3DeploymentJson['native-token'],
+      };
+    } catch {
+      // empty on purpose
+    }
+  }
+
+  throw new Error('nitro-testnode sequencer not found');
 }
